Distribute auxiliary functions among modules
