+++ web/app/countries_structure.py
@@ -1,5 +1,4 @@
 import itertools
-from queue import Queue
 
 from config import (
     TABLE as table,
@@ -385,104 +384,3 @@ def _get_country_osm_id_by_name(conn, name):
     if not rec:
         raise CountryStructureException(f'Not found country "{name}"')
     return int(rec[0])
-
-
-def is_administrative_region(conn, region_id):
-    cursor = conn.cursor()
-    cursor.execute(f"""
-        SELECT count(1) FROM {osm_table} WHERE osm_id = %s
-        """, (region_id,)
-    )
-    count = cursor.fetchone()[0]
-    return (count > 0)
-
-
-def find_osm_child_regions(conn, region_id):
-    cursor = conn.cursor()
-    cursor.execute(f"""
-        SELECT c.id, oc.admin_level
-        FROM {table} c, {table} p, {osm_table} oc
-        WHERE p.id = c.parent_id AND c.id = oc.osm_id
-            AND p.id = %s
-        """, (region_id,)
-    )
-    children = []
-    for rec in cursor:
-        children.append({'id': int(rec[0]), 'admin_level': int(rec[1])})
-    return children
-
-
-def is_leaf(conn, region_id):
-    cursor = conn.cursor()
-    cursor.execute(f"""
-        SELECT count(1)
-        FROM {table}
-        WHERE parent_id = %s
-        """, (region_id,)
-    )
-    count = cursor.fetchone()[0]
-    return (count == 0)
-
-
-def get_region_country(conn, region_id):
-    """Returns the uppermost predecessor of the region in the hierarchy,
-    possibly itself.
-    """
-    predecessors = get_predecessors(conn, region_id)
-    return predecessors[-1]
-
-
-def get_predecessors(conn, region_id):
-    """Returns the list of (id, name)-tuples of all predecessors,
-    starting from the very region_id.
-    """
-    predecessors = []
-    cursor = conn.cursor()
-    while True:
-        cursor.execute(f"""
-            SELECT id, name, parent_id
-            FROM {table} WHERE id={region_id}
-            """
-        )
-        rec = cursor.fetchone()
-        if not rec:
-           raise Exception(f"No record in '{table}' table with id = {region_id}")
-        predecessors.append(rec[0:2])
-        parent_id = rec[2]
-        if not parent_id:
-            break
-        region_id = parent_id
-    return predecessors
-
-
-def get_region_full_name(conn, region_id):
-    predecessors = get_predecessors(conn, region_id)
-    return '_'.join(pr[1] for pr in reversed(predecessors))
-
-
-def get_similar_regions(conn, region_id, only_leaves=False):
-    """Returns ids of regions of the same admin_level in the same country.
-    Prerequisite: is_administrative_region(region_id) is True.
-    """
-    cursor = conn.cursor()
-    cursor.execute(f"""
-        SELECT admin_level FROM {osm_table}
-        WHERE osm_id = %s""", (region_id,)
-    )
-    admin_level = int(cursor.fetchone()[0])
-    country_id, country_name = get_region_country(conn, region_id)
-    q = Queue()
-    q.put({'id': country_id, 'admin_level': 2})
-    similar_region_ids = []
-    while not q.empty():
-        item = q.get()
-        if item['admin_level'] == admin_level:
-            similar_region_ids.append(item['id'])
-        elif item['admin_level'] < admin_level:
-            children = find_osm_child_regions(item['id'])
-            for ch in children:
-                q.put(ch)
-    if only_leaves:
-        similar_region_ids = [r_id for r_id in similar_region_ids
-                                  if is_leaf(conn, r_id)]
-    return similar_region_ids
